requirements.txt
streamlit==1.31.1
requests==2.31.0
pymongo==4.6.1
python-dotenv==1.0.0